// src/models/user.ts
export default interface User {
    _id: string;
    user_name: string;
    user_status: string;
    user_status_string: string;
    reward_id: string;
    color_palette: string;
    render_in_front: boolean;
    name: string;
    follower_count: number;
    level: string;
    tier: string;
    user_rank: string;
    coffee: number;
    pfp_path: string;
}